Create Data Exploration with SQL for COVID-19 Data
This project is focused on data exploration methods using Microsoft SQL Server Management Studio to describe the main characteristics of a data set. In this case, I explored the specifics of the data within a dataset on the COVID-19 pandemic from Our World in Data.org. This dataset tracked the total number of deaths between years 2020 and 2021 caused by COVID-19 based on “location”, “date”, “population”, “total cases”, “new cases” and “total deaths”. The other aspect of this dataset that I explored is data on the COVID-19 vaccinations based on continent, location, date, new tests, total tests administered per case and the overall positive rate per test. The question I was seeking to answer when exploring the data for these large datasets is how COVID-19 affected people worldwide during the pandemic and what was the response to mitigate any further harm that was caused to people around the globe due to the virus. Specifically, I wanted to understand the overall impact that COVID-19 had on the United States. The SQL skills I used to properly explore the data within this project are Joins, CTE's, Temp Tables, Windows Functions, Aggregate Functions, Creating Views, Converting Data Types.