Create Exploratory Data Analysis of popular Apple App Store Apps using SQLite
This project is intended to perform exploratory data analysis and data cleaning using the relational database programming language, SQL. The SQL software that I used in this project is SQLite. The source of the data for this project are two datasets found on Kaggle.com. appleStore.CSV is the first file containing the dataset about Apple apps from the Apple Store based on app names, the size of the app in term of bytes, devices that can operate the app and languages offered for the app. The second file, appleStore_description.csv, contains the other dataset which provides specific information about how the capabilities and purpose of each app can benefit its userbase and its overall perception within the app store marketplace.

Stakeholder Questions about Apple App Store Data:  
	
	In this project I play a role as app developer that has interest in the analysis and findings of the data within this project that needs data driven insights to decide what type of app to build to be competitive within the Apple App Store. To make the best app possible, I had to asked the following questions about app data within the Apple App Store. 

•	What app categories are most popular?
•	What prices should I set?
•	How can I maximize user ratings?

After analyzing the most popular apps within the Apple App store  here are the final recommendations for me to develop a unique and popular app and the answers to the questions I wanted to know about the data:

1.	Paid apps have better ratings overall.

2.	Apps supporting between 10 and 30 languages have better ratings.

3.	Finance and Book apps have the lowest ratings within the Apple App Store

4.	Apps with longer descriptions on average have better ratings than those with shorter descriptions.

5.	I should aim to create an app that has an average user rating above 3.5 to be competitive in app marketplace.

6.	Games and Entertainment apps have the most competition with the Apple App Store because the market is saturated with these apps. So I must make a unique app within these genres to separate from the competition or develop an app in an entirely different genre.