Add files via upload
This project is from a guided course that I completed through Coursera.org. In this guided project I utilized Microsoft Excel to analyze sales data from a sample company. To sort and filter the data within the project I used filtering and sorting tools within Excel to access specific information about the data. In order to create new data and relate data from different tables I used the IF and VLOOKUP functions available in Excel. The final tool I used to further analyze data within the dataset was PivotTables to be able to clearly summarize and make comparisons with the data.