Add files via upload
This is a guided project from a Microsoft Excel course on Coursera.org. In this project, I learned how to analyze data and identify trends for employee sales data using a variety of tools in Microsoft Excel. To highlight specific sales data such as percent progress of sales by quarter fiscal year and total sales revenue I used Conditional Formatting and charts in order to represent the data in visual form. I then utilized PivotTables to create summaries of the sales data that focused on total sales revenue by location division such as East, West, North and South and the sum of each quarter fiscal year. I further represented the data in this project visually using PivotTables to create a column chart showing the percentage of sales revenue progress overtime for all the sales people.