Add files via upload
This guided project that I completed was from a Microsoft Excel course on Coursera.org. In this project I utilized formulas and functions to analyze sales data from a fictitious company that sells different models and styles of chairs. While completing this project, formulas of addition, subtraction, multiplication and division were required in order to calculate total profits of sales data to determine if profits were being made by the company. Popularity of items sold, the averages and monthly percentages of profits were essential to analyzing the data accurately and to determine the numerical values of these sales factors I had to use SUM and SUMIF functions to ensure the company was successful in there operations in selling chairs.